+++ brut-components/src/main/java/org/bloomreach/forge/brut/components/annotation/ComponentTestConfig.java
@@ -1,0 +1,20 @@
+package org.bloomreach.forge.brut.components.annotation;
+
+class ComponentTestConfig {
+
+    private final String annotatedClassesResourcePath;
+    private final String testResourcePath;
+
+    ComponentTestConfig(String annotatedClassesResourcePath, String testResourcePath) {
+        this.annotatedClassesResourcePath = annotatedClassesResourcePath;
+        this.testResourcePath = testResourcePath;
+    }
+
+    String getAnnotatedClassesResourcePath() {
+        return annotatedClassesResourcePath;
+    }
+
+    String getTestResourcePath() {
+        return testResourcePath;
+    }
+}